Move code in preparation
1. Collision filter legacy into its own file.
2. Shape-shape infrastructure into distance_to_shape.h. The move included
   two small changes:
  - DistanceData and DistanceCallback have been renamed to CallbackData
    and Callback (in the shape_distance namespace) to match the pattern
    used in distance_to_point.h)
  - Interfaces overly inspired by FCL's habit of passing pointers have
    been updated to Drake-standard const&.
3. Clean up of headers and aliases in proximity_engine.cc
