+++ geometry/proximity/BUILD.bazel
@@ -13,9 +13,21 @@ package(default_visibility = ["//visibility:public"])
 drake_cc_package_library(
     name = "proximity",
     deps = [
+        ":collision_filter_legacy",
         ":distance_to_point",
         ":distance_to_point_with_gradient",
+        ":distance_to_shape",
         ":proximity_utilities",
+    ],
+)
+
+drake_cc_library(
+    name = "collision_filter_legacy",
+    hdrs = ["collision_filter_legacy.h"],
+    deps = [
+        ":proximity_utilities",
+        "//common:essential",
+        "//common:sorted_vectors_have_intersection",
     ],
 )
 
@@ -43,6 +55,17 @@ drake_cc_library(
         "//geometry/query_results:signed_distance_to_point",
         "//math:geometric_transform",
         "@fcl",
+    ],
+)
+
+drake_cc_library(
+    name = "distance_to_shape",
+    hdrs = ["distance_to_shape.h"],
+    deps = [
+        ":collision_filter_legacy",
+        ":distance_to_point",
+        ":proximity_utilities",
+        "//geometry/query_results:signed_distance_pair",
     ],
 )
